Disable v6 NURBS to pass regression tests.
v6 NURBS use a different chunk for the NURBS formula and the assumption that the formula was in the NURBSTo chunk was causing an infinite loop.

+++ src/lib/VSDXContentCollector.cpp
@@ -613,7 +613,7 @@ void libvisio::VSDXContentCollector::collectNURBSTo(unsigned id, unsigned level,
   transformPoint(m_x, m_y);
 }
 
-double libvisio::VSDXContentCollector::_NURBSBasis(unsigned knot, unsigned degree, double point, double controlCount, const std::vector<double> &knotVector)
+double libvisio::VSDXContentCollector::_NURBSBasis(unsigned knot, unsigned degree, double point, unsigned controlCount, const std::vector<double> &knotVector)
 {
   if (degree == 0)
   {
