+++ src/lib/VSDXContentCollector.h
@@ -81,7 +81,7 @@ private:
   void transformPoint(double &x, double &y);
   void transformAngle(double &angle);
 
-  double _NURBSBasis(unsigned knot, unsigned degree, double point, double controlCount, const std::vector<double> &knotVector);
+  double _NURBSBasis(unsigned knot, unsigned degree, double point, unsigned controlCount, const std::vector<double> &knotVector);
 
   void _flushCurrentPath();
   void _flushCurrentForeignData();
